feat: use packaged apps (preliminary)
- add package extraction functions to app files model
- adapt ds-host frontend, routes to accept package file instead of dir
- ds-dev can now be used with packaged app

+++ cmd/ds-host/models/appfilesmodel/appfilesmodel.go
@@ -1,9 +1,12 @@
 package appfilesmodel
 
 import (
+	"archive/tar"
+	"compress/gzip"
 	"encoding/json"
 	"errors"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -25,68 +28,125 @@ type AppFilesModel struct {
 	Config *domain.RuntimeConfig `checkinject:"required"`
 }
 
-// Save puts the data passed in files in an apps directory
-// Wonder if we will need to save additional files here if for ex the manifest points to a file as URL?
-func (a *AppFilesModel) Save(files *map[string][]byte) (string, error) {
-	logger := a.getLogger("Save()")
-	appsPath := a.Config.Exec.AppsPath
+// SavePackage creates a location key and saves the package file under /package
+// because we want to save it as-is for later download.
+func (a *AppFilesModel) SavePackage(r io.Reader) (string, error) { // this should be areader not bytes
+	logger := a.getLogger("SavePackage()")
 
-	err := os.MkdirAll(appsPath, 0766)
+	locationKey, err := a.createLocation()
 	if err != nil {
-		logger.AddNote("os.MkdirAll() appsPath").Error(err)
-		return "", errors.New("internal error saving app files")
+		return "", err
 	}
 
-	appPath, err := ioutil.TempDir(appsPath, "app")
+	f, err := os.OpenFile(a.getPackagePath(locationKey), os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0644)
 	if err != nil {
-		logger.AddNote("ioutil.TempDir()").Error(err)
-		return "", errors.New("internal error saving app files")
+		logger.AddNote("OpenFile").Error(err)
+		return locationKey, err
 	}
-
-	locationKey := filepath.Base(appPath)
-
-	appPath = filepath.Join(appPath, "app")
-	err = os.MkdirAll(appPath, 0766) // omg permissions!
+	defer f.Close()
+	_, err = io.Copy(f, r)
 	if err != nil {
-		logger.AddNote("os.MkdirAll() app").Error(err)
-		return "", errors.New("internal error saving app files")
-	}
-
-	logger.AddNote("files loop")
-	writeErr := false
-	for f, data := range *files {
-		fPath := filepath.Join(appPath, f)
-		inside, err := pathInsidePath(fPath, appPath)
-		if err != nil {
-			logger.AddNote("pathInsidePath()").Error(err)
-			writeErr = true
-			break
-		}
-		if !inside {
-			logger.Log(fmt.Sprintf("file path outside of app path: %v", f))
-			writeErr = true
-			break
-		}
-
-		err = os.MkdirAll(filepath.Dir(fPath), 0766)
-		if err != nil {
-			logger.AddNote(fmt.Sprintf("os.MkdirAll(): %v", f)).Error(err)
-			return "", errors.New("internal error saving app files")
-		}
-
-		err = ioutil.WriteFile(fPath, data, 0666) // TODO: correct permissions?
-		if err != nil {
-			logger.AddNote(fmt.Sprintf("ioutil.WriteFile(): %v", f)).Error(err)
-			writeErr = true
-			break
-		}
-	}
-
-	if writeErr {
-		return "", errors.New("internal error saving app files")
+		logger.AddNote("io.Copy").Error(err)
+		return locationKey, fmt.Errorf("internal error saving package: %w", err)
 	}
 
 	return locationKey, nil
+}
+
+func (a *AppFilesModel) createLocation() (string, error) {
+	p := a.Config.Exec.AppsPath
+
+	err := os.MkdirAll(p, 0766)
+	if err != nil {
+		a.getLogger("createLocation()").AddNote("os.MkdirAll() apps").Error(err)
+		return "", fmt.Errorf("internal error creating location: %w", err)
+	}
+
+	p, err = ioutil.TempDir(p, "app")
+	if err != nil {
+		a.getLogger("createLocation()").AddNote("ioutil.TempDir()").Error(err)
+		return "", fmt.Errorf("internal error creating location: %w", err)
+	}
+
+	// maybe assert that the L2P returns the path we created?
+
+	return filepath.Base(p), nil
+}
+
+// ExtractPackage expands the package contents.
+func (a *AppFilesModel) ExtractPackage(locationKey string) error {
+	// should ensure that the directory is empty before?
+
+	logger := a.getLogger("SavePackage()")
+
+	packageFD, err := os.Open(a.getPackagePath(locationKey))
+	if err != nil {
+		logger.AddNote("os.Open package").Error(err)
+		return err
+	}
+	defer packageFD.Close()
+
+	appFilesPath := a.AppLocation2Path.Files(locationKey)
+	err = os.MkdirAll(appFilesPath, 0766)
+	if err != nil {
+		logger.AddNote("os.MkDirAll appFilesPath").Error(err)
+		return err
+	}
+
+	err = ExtractPackageLow(packageFD, appFilesPath)
+	if err != nil {
+		logger.AddNote("extractPackage").Error(err)
+		return err
+	}
+	return nil
+}
+
+func ExtractPackageLow(r io.Reader, appFilesPath string) error {
+	gzr, err := gzip.NewReader(r)
+	if err != nil {
+		return err
+	}
+	defer gzr.Close()
+
+	tr := tar.NewReader(gzr)
+	for {
+		hdr, err := tr.Next()
+		if err == io.EOF {
+			break // End of archive
+		}
+		if err != nil {
+			return err
+		}
+		name := filepath.Clean(hdr.Name)
+		if filepath.IsAbs(name) {
+			return errors.New("invalid (absolute) path in tar header: " + name)
+		}
+		if strings.Contains(name, "..") {
+			return errors.New("invalid (outside) path in tar header: " + name)
+		}
+		p := filepath.Join(appFilesPath, name)
+		err = os.MkdirAll(filepath.Dir(p), 0766)
+		if err != nil {
+			return fmt.Errorf("failed to create dir: %s: %w", filepath.Dir(p), err)
+		}
+		fd, err := os.OpenFile(p, os.O_WRONLY|os.O_CREATE|os.O_EXCL, 0644)
+		if os.IsExist(err) {
+			return errors.New("file already exists: " + name)
+		}
+		if err != nil {
+			return fmt.Errorf("failed to create file: %s: %w", name, err)
+		}
+		_, err = io.Copy(fd, tr)
+		fd.Close()
+		if err != nil {
+			return fmt.Errorf("failed to write file: %s: %w", name, err)
+		}
+	}
+	return nil
+}
+
+func (a *AppFilesModel) getPackagePath(locationKey string) string {
+	return filepath.Join(a.AppLocation2Path.Base(locationKey), "package.tar.gz")
 }
 
 // ReadManifest reads metadata from the files at location key
@@ -110,6 +170,7 @@ func (a *AppFilesModel) ReadManifest(locationKey string) (domain.AppVersionManif
 
 	meta, err := unmarshalManifest(jsonBytes)
 	if err != nil {
+		a.getLogger(fmt.Sprintf("ReadManifest(), location key: %v, unmarshalManifest", locationKey)).Error(err)
 		return domain.AppVersionManifest{}, err
 	}
 
@@ -118,7 +179,7 @@ func (a *AppFilesModel) ReadManifest(locationKey string) (domain.AppVersionManif
 
 func (a *AppFilesModel) WriteRoutes(locationKey string, routesData []byte) error {
 	routesFile := filepath.Join(a.AppLocation2Path.Meta(locationKey), "routes.json")
-	err := ioutil.WriteFile(routesFile, routesData, 0666) // TODO: correct permissions?
+	err := ioutil.WriteFile(routesFile, routesData, 0644) // TODO: correct permissions?
 	if err != nil {
 		a.getLogger(fmt.Sprintf("WriteRoutes(), location key: %v", locationKey)).Error(err)
 		return err
@@ -160,10 +221,14 @@ func (a *AppFilesModel) ReadMigrations(locationKey string) ([]byte, error) {
 
 // Delete removes the files from the system
 func (a *AppFilesModel) Delete(locationKey string) error {
+	if locationKey == "" {
+		e := errors.New("empty string is not a valid location key")
+		a.getLogger("Delete()").Error(e)
+		return e
+	}
 	if !a.locationKeyExists(locationKey) {
 		return nil //is that an error or do we consider this OK?
 	}
-
 	err := os.RemoveAll(a.AppLocation2Path.Base(locationKey))
 	if err != nil {
 		a.getLogger("Delete()").Error(err)
